+++ CHANGELOG.md
@@ -26,6 +26,8 @@
 - [usd#2017](https://github.com/Autodesk/arnold-usd/issues/2017) - Set root transform and node id in the hydra procedural
 - [usd#2015](https://github.com/Autodesk/arnold-usd/issues/2015) - Support hydra points with empty widths
 - [usd#2008](https://github.com/Autodesk/arnold-usd/issues/2008) - Write spot and photometric lights as UsdLux schemas
+- [usd#2030](https://github.com/Autodesk/arnold-usd/issues/2030) - Write background and atmosphere shaders under a node graph for hydra support
+- [usd#2031](https://github.com/Autodesk/arnold-usd/issues/2031) - Write AOV shaders under a node graph for hydra support
 - [usd#2025](https://github.com/Autodesk/arnold-usd/issues/2025) - Write imagers through node graphs for hydra support
 
 ### Bug fixes